Refactor out conditional imports
Place conditional imports used by the various tests in to a single module,
`tests/common.py`. Reference this module instead of placing nearly identical
conditional imports in to each individual test module.

This change makes the imports in each test module more concise. However, this
change also introduces a maintenance hassle. Linters are not able to check and
see whether an import is imported in another module, so it's quite possible for
unused imports to start piling up in the `tests.common` module.