WIP: Change procedure for finding entity IDs
The server may return JSON data in the following forms for a `OneToOneField`:

    'user': None
    'user': {'name': 'Alice Hayes', 'login': 'ahayes', 'id': 1}
    'user_id': 1
    'user_id': None

The algorithm that looks for these IDs first checks for the presence of
an embedded object, like the first two lines above, and then proceeds to
look for a plain ID, like the second two lines above. This is
problematic in cases where both an embedded object and a plain ID are
present, like this:

    "results": [
        "sync_plan_id": 1,
        "sync_plan": {
            "sync_date": "1900-01-01 02:00:00 UTC",
            "next_sync": "2016-05-03 02:00:00 UTC",
            "interval": "daily",
            "name": "Nightly Sync",
            "description": null
        },
    ]

The algorithm is problematic because it will decide to fetch the ID at
`results[0]['sync_plan']['id']`, and this ID does not exist. Fix this
issue by reversing the order in which the algorithm searches for IDs and
looking for plain IDs before embedded objects.